Remove build.ts
