+++ package.json
@@ -107,6 +107,7 @@
         "typescript": "4.6.2",
         "url-loader": "4.1.1",
         "webpack": "4.46.0",
+        "webpack-cli": "4.9.2",
         "winston": "3.6.0"
     },
     "devDependencies": {
